refactor: self-study 필터 로직을 toggleFilter·리듀서·훅으로 분리
toggleFilter를 shared/lib으로 이동하고, 상태 머신(studyFilterReducer)과
useStudyFilter 훅을 독립 파일로 분리해 SelfStudySection을 UI 전용으로 정리
