+++ src/features/self-study/model/studyFilterReducer.ts
@@ -1,11 +1,5 @@
-import { useReducer } from "react";
-import { getClassNumber, getGrade } from "@/entities/user/lib/getUserInfo";
+import { toggleFilter } from "@/shared/lib/toggleFilter";
 import type { Sex } from "@/entities/user/model/user";
-import type { DormitoryStudent } from "@/entities/dormitory/model/types";
-
-export function toggleFilter<T>(arr: T[], value: T): T[] {
-  return arr.includes(value) ? arr.filter((v) => v !== value) : [...arr, value];
-}
 
 export interface FilterState {
   searchQuery: string;
@@ -50,33 +44,4 @@ export function filterReducer(
     case "RESET":
       return initialFilterState;
   }
-}
-
-export function useStudyFilter(students: DormitoryStudent[]) {
-  const [state, dispatch] = useReducer(filterReducer, initialFilterState);
-  const { searchQuery, selectedGrades, selectedClasses, selectedGender } =
-    state;
-
-  const filteredStudents = students.filter((s) => {
-    if (
-      searchQuery &&
-      !s.name.includes(searchQuery) &&
-      !String(s.studentNumber).includes(searchQuery)
-    )
-      return false;
-    if (
-      selectedGrades.length > 0 &&
-      !selectedGrades.includes(getGrade(s.studentNumber))
-    )
-      return false;
-    if (
-      selectedClasses.length > 0 &&
-      !selectedClasses.includes(getClassNumber(s.studentNumber))
-    )
-      return false;
-    if (selectedGender && s.sex !== selectedGender) return false;
-    return true;
-  });
-
-  return { state, filteredStudents, dispatch };
 }
