+++ src/shared/lib/toggleFilter.ts
@@ -1,0 +1,3 @@
+export function toggleFilter<T>(arr: T[], value: T): T[] {
+  return arr.includes(value) ? arr.filter((v) => v !== value) : [...arr, value];
+}
